Refactoring logic
