+++ src/gif.py
@@ -1,46 +1,27 @@
 from PIL import Image
 import time
-
-NUM_LEDS = 12 * 11
-
-
-def cartesian_to_neopixel(x, y, width=12):
-    if y % 2 == 0:
-        row_index = NUM_LEDS - ((y) * width)
-        index = row_index - (x + 1)
-    else:
-        row_index = NUM_LEDS - ((y + 1) * width)
-        index = row_index + x
-    return index
+from clock_display_hal import ClockDisplayHAL
 
 
-def set_pixel(pixels, x, y, color, width=12):
-    index = cartesian_to_neopixel(x, y, width)
-    pixels[index] = color
-
-
-def update_led_pixels(led_pixels, gif_pixels, new_size):
+def update_led_pixels(gif_pixels, new_size, clock_display_hal):
     for y in range(new_size[1]):
         for x in range(new_size[0]):
             r, g, b = gif_pixels[x, y]
-            set_pixel(led_pixels, x, y, (r, g, b))
+            clock_display_hal.set_pixel(x, y, (r, g, b))
 
 
-def display_gif(led_pixels):
-    gif_path = "../heart-art.gif"
+def display_gif(gif_path, clock_display_hal, display_gif_duration=4):
+    gif_path = gif_path
     img = Image.open(gif_path)
-    new_size = (12, 11)
-
-    display_time_of_gif = 4
+    new_size = (ClockDisplayHAL.WIDTH, ClockDisplayHAL.HEIGHT)
     start_time = time.time()
-
-    while time.time() < start_time + display_time_of_gif:
+    while time.time() < start_time + display_gif_duration:
         try:
             frame = img.resize(new_size).convert('RGB')
             gif_pixels = frame.load()
 
-            update_led_pixels(led_pixels, gif_pixels, new_size)
-            led_pixels.show()
+            update_led_pixels(gif_pixels, new_size, clock_display_hal)
+            clock_display_hal.show()
             time.sleep(0.1)
             img.seek(img.tell() + 1)
